:sparkles: Add the flow detail page.
